+++ prisma/schema.prisma
@@ -1,11 +1,17 @@
-// This is your Prisma schema file,
-// learn more about it in the docs: https://pris.ly/d/prisma-schema
+
+datasource db {
+  provider = "postgresql"
+  url      = env("DATABASE_URL")
+}
+
 
 generator client {
   provider = "prisma-client-js"
 }
 
-datasource db {
-  provider = "postgresql"
-  url      = env("DATABASE_URL")
+model Programmer {
+    id Int @id @default(autoincrement())
+    name String 
+    knownFor String 
+    vote Int
 }
