--- schema.prisma
@@ -1,17 +1,0 @@
-
-datasource db {
-  provider = "postgresql"
-  url      = env("DATABASE_URL")
-}
-
-
-generator client {
-  provider = "prisma-client-js"
-}
-
-model Programmer {
-    id Int @id @default(autoincrement())
-    name String 
-    knownFor String 
-    vote Int
-}
